backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.2.0
gunicorn==21.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
googletrans==4.0.0rc1
